Decouple Metrics class from ES classes
