+++ Source/santad/EventProviders/SNTEndpointSecurityTreeAwareClient.h
@@ -14,15 +14,15 @@
 
 #include "Source/santad/EventProviders/SNTEndpointSecurityClient.h"
 
+#include "Source/common/es/ESMetricsObserver.h"
 #include "Source/santad/EventProviders/EndpointSecurity/EndpointSecurityAPI.h"
-#include "Source/santad/Metrics.h"
 #include "Source/santad/ProcessTree/process_tree.h"
 
 @interface SNTEndpointSecurityTreeAwareClient : SNTEndpointSecurityClient
 @property std::shared_ptr<santa::santad::process_tree::ProcessTree> processTree;
 
 - (instancetype)initWithESAPI:(std::shared_ptr<santa::EndpointSecurityAPI>)esApi
-                      metrics:(std::shared_ptr<santa::Metrics>)metrics
+                      metrics:(std::shared_ptr<santa::ESMetricsObserver>)metrics
                     processor:(santa::Processor)processor
                   processTree:
                       (std::shared_ptr<santa::santad::process_tree::ProcessTree>)processTree;
